+++ packages/ubcli/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `ubcli meta-tr` will transform entity level mapping to object (missed in prev. version)
+ - `ubcli meta-tr` will fix mapping dialect AnsiSql -> AnsiSQL (many such mistakes in old UB projects)
 
 ### Changed